refactor: barrel export を制限
コンポーネント定義以外の index.ts 再エクスポートを削除し、oxc/no-barrel-file で export * を禁止する。

test-utils の Testing Library 再公開は例外として設定する。
