+++ apps/web/src/app/Header/index.ts
@@ -1,1 +1,1 @@
-export * from "./Header"
+export { Header } from "./Header"
