+++ apps/web/src/app/Provider.tsx
@@ -2,8 +2,8 @@ import { QueryClientProvider } from "@tanstack/react-query"
 import type { ReactNode } from "react"
 
 import { createQueryClient } from "../lib/queryClient"
-import { SnackbarProvider } from "../providers/snackbar"
-import { SupabaseSessionProvider } from "../providers/supabase"
+import { SnackbarProvider } from "../providers/snackbar/SnackbarProvider"
+import { SupabaseSessionProvider } from "../providers/supabase/SupabaseSessionProvider"
 import { ThemeProvider } from "../providers/theme/ThemeProvider"
 
 const queryClient = createQueryClient()
