+++ apps/web/src/app/Sidebar/index.ts
@@ -1,2 +1,1 @@
 export { Sidebar } from "./Sidebar"
-export { useSidebar } from "./useSidebar"
