+++ apps/web/src/app/routes/AggregatesPage/index.ts
@@ -1,1 +1,1 @@
-export * from "./AggregatesPage"
+export { AggregatesPage } from "./AggregatesPage"
