+++ apps/web/src/app/routes/PaymentsPage/index.ts
@@ -1,1 +1,1 @@
-export * from "./PaymentsPage"
+export { PaymentsPage } from "./PaymentsPage"
