+++ apps/web/src/app/routes/SettingsPage/index.ts
@@ -1,1 +1,1 @@
-export * from "./SettingsPage"
+export { SettingsPage } from "./SettingsPage"
