+++ apps/web/src/components/inputs/AmountInput/index.ts
@@ -1,1 +1,1 @@
-export * from "./AmountInput"
+export { AmountInput } from "./AmountInput"
