+++ apps/web/src/components/misc/Accordion/index.ts
@@ -1,1 +1,1 @@
-export * from "./Accordion"
+export { Accordion, AccordionItem, AccordionTrigger, AccordionContent } from "./Accordion"
