+++ apps/web/src/components/misc/Logo/index.ts
@@ -1,1 +1,1 @@
-export * from "./Logo"
+export { Logo } from "./Logo"
