--- apps/web/src/features/budgets/createMonthlyBudget/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreateMonthlyBudgetModal"
