--- apps/web/src/features/budgets/latestMonthlyBudget/index.ts
@@ -1,1 +1,0 @@
-export { LatestMonthlyBudget } from "./LatestMonthlyBudget/LatestMonthlyBudget"
