--- apps/web/src/features/budgets/listCategoryBudget/index.ts
@@ -1,2 +1,0 @@
-export * from "./fetchCategoryBudgets"
-export * from "./useCategoryBudgets"
