--- apps/web/src/features/budgets/listMonthlyBudget/index.ts
@@ -1,1 +1,0 @@
-export * from "./MonthlyBudgetList/MonthlyBudgetList"
