+++ apps/web/src/features/categories/components/CategorySelect/index.ts
@@ -1,1 +1,7 @@
-export * from "./CategorySelect"
+export {
+  CategorySelect,
+  CategoryOption,
+  NoneCategoryOption,
+  ErrorCategoryOption,
+  LoadingCategoryOption,
+} from "./CategorySelect"
