+++ apps/web/src/features/payments/components/AmountInput/index.ts
@@ -1,1 +1,1 @@
-export * from "./AmountInput"
+export { AmountInput } from "./AmountInput"
