+++ apps/web/src/features/payments/components/NoteInput/index.ts
@@ -1,1 +1,1 @@
-export * from "./NoteInput"
+export { NoteInput } from "./NoteInput"
