+++ apps/web/src/features/payments/createPayment/CreatePaymentModal/index.ts
@@ -1,1 +1,1 @@
-export * from "./CreatePaymentModal"
+export { CreatePaymentModal } from "./CreatePaymentModal"
