--- apps/web/src/features/payments/createPayment/index.ts
@@ -1,1 +1,0 @@
-export * from "./CreatePaymentModal"
