--- apps/web/src/features/payments/deletePayment/index.ts
@@ -1,1 +1,0 @@
-export { useDeletePaymentModal } from "./useDeletePaymentModal"
