+++ apps/web/src/features/payments/listPayment/PaymentCategoryFilter/index.ts
@@ -1,1 +1,1 @@
-export * from "./PaymentCategoryFilter"
+export { PaymentCategoryFilter } from "./PaymentCategoryFilter"
