+++ apps/web/src/features/payments/listPayment/PaymentList/PaymentList.tsx
@@ -8,7 +8,8 @@ import type { Payment, PaymentId } from "../../../../types/payment"
 import { getCategoryStrict, toCategoryMap } from "../../../categories/listCategory/toCategoryMap"
 import { useCategories } from "../../../categories/listCategory/useCategories"
 import { DeletePaymentModal } from "../../deletePayment/DeletePaymentModal"
-import { PaymentDetailsOverlay, usePaymentDetailsState } from "../../paymentDetails"
+import { PaymentDetailsOverlay } from "../../paymentDetails/PaymentDetailsOverlay"
+import { usePaymentDetailsState } from "../../paymentDetails/usePaymentDetailsState"
 import { PaymentItem } from "../PaymentItem"
 import { useCategoryId } from "../useCategoryId"
 import { usePayments } from "../usePayments"
