+++ apps/web/src/features/payments/listPayment/PaymentList/index.ts
@@ -1,1 +1,1 @@
-export * from "./PaymentList"
+export { PaymentList } from "./PaymentList"
