--- apps/web/src/features/payments/listPayment/index.ts
@@ -1,2 +1,0 @@
-export * from "./PaymentCategoryFilter"
-export * from "./PaymentList"
