+++ apps/web/src/features/payments/paymentDetails/AmountField/AmountField.stories.tsx
@@ -3,7 +3,7 @@ import { QueryClientProvider } from "@tanstack/react-query"
 import { fn } from "storybook/test"
 
 import { createQueryClient } from "../../../../lib/queryClient"
-import { SnackbarProvider } from "../../../../providers/snackbar"
+import { SnackbarProvider } from "../../../../providers/snackbar/SnackbarProvider"
 import { ThemeProvider } from "../../../../providers/theme/ThemeProvider"
 import { createPaymentHandlers } from "../../../../test/msw/handlers/payments"
 import { AmountField } from "./AmountField"
