+++ apps/web/src/features/payments/paymentDetails/AmountField/index.ts
@@ -1,1 +1,1 @@
-export * from "./AmountField"
+export { AmountField } from "./AmountField"
