+++ apps/web/src/features/payments/paymentDetails/CategoryField/index.ts
@@ -1,1 +1,1 @@
-export * from "./CategoryField"
+export { CategoryField } from "./CategoryField"
