+++ apps/web/src/features/payments/paymentDetails/EditableField/index.ts
@@ -1,1 +1,1 @@
-export * from "./EditableField"
+export { EditableField } from "./EditableField"
