+++ apps/web/src/features/payments/paymentDetails/NoteField/index.ts
@@ -1,1 +1,1 @@
-export * from "./NoteField"
+export { NoteField } from "./NoteField"
