+++ apps/web/src/features/payments/paymentDetails/PaymentDateField/index.ts
@@ -1,1 +1,1 @@
-export * from "./PaymentDateField"
+export { PaymentDateField } from "./PaymentDateField"
