+++ apps/web/src/features/payments/paymentDetails/SubmitIconButton/index.ts
@@ -1,1 +1,1 @@
-export * from "./SubmitIconButton"
+export { SubmitIconButton } from "./SubmitIconButton"
