--- apps/web/src/features/payments/paymentDetails/index.ts
@@ -1,3 +1,0 @@
-export * from "./PaymentDetailsOverlay"
-export { usePaymentDetails } from "./usePaymentDetails"
-export { usePaymentDetailsState } from "./usePaymentDetailsState"
