--- apps/web/src/features/payments/updatePayment/index.ts
@@ -1,1 +1,0 @@
-export { useUpdatePayment } from "./useUpdatePayment"
