+++ apps/web/src/features/summaryByMonth/CategoryTotals/index.ts
@@ -1,1 +1,1 @@
-export * from "./CategoryTotals"
+export { CategoryTotals } from "./CategoryTotals"
