--- apps/web/src/features/summaryByMonth/index.ts
@@ -1,2 +1,0 @@
-export { MonthSelector } from "./MonthSelector"
-export { Summary } from "./Summary"
