--- apps/web/src/providers/snackbar/index.ts
@@ -1,1 +1,0 @@
-export { SnackbarProvider, useSnackbar } from "./SnackbarProvider"
