--- apps/web/src/providers/supabase/index.ts
@@ -1,2 +1,0 @@
-export { SupabaseSessionProvider } from "./SupabaseSessionProvider"
-export { useSupabaseSession } from "./useSupabaseSession"
